+++ src/bot/commands/ping.ts
