+++ src/bot/commands/reset.ts
