+++ src/bot/commands/setup.ts
