+++ src/bot/database/entities/User.ts
