+++ src/bot/hs_auth/client.ts
